requirements.txt
requests
beautifulsoup4
selenium
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
schedule
pypdf
fake-useragent
webdriver-manager
loguru
pytest
